new commit

+++ app.py
@@ -4,10 +4,14 @@ import os
 from datetime import datetime
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user, current_user
-from webforms import ProductForm, UserForm, LoginForm
+from webforms import ProductForm, UserForm, LoginForm, SearchForm
+from flask_ckeditor import CKEditor
+
 
 app = Flask(__name__)
 
+#rich text editor init
+ckeditor = CKEditor(app)
 
 env_config = os.getenv("APP_SETTINGS", "config.DevelopmentConfig")
 app.config.from_object(env_config)
@@ -105,45 +109,57 @@ def user_form():
     users=db.session.query(Users).all()
     return render_template('createuser.html', form=form, users=users)
 
-@app.route('/updateuser/<int:id>', methods=['GET','POST'])
+@app.route('/updateuser/<int:id>', methods=['GET','POST'])   #methods=["PATCH"]
 @login_required
 def update_user(id):
     form = UserForm()
     thing_to_update= Users.query.get_or_404(id)
-    if request.method =='POST':
-        thing_to_update.username=request.form['username']
-        thing_to_update.email=request.form['email']
-        thing_to_update.password_hash = generate_password_hash(form.password_hash.data, "sha256")
-        thing_to_update.address = request.form['address']
-        thing_to_update.postal_code = request.form['postal_code']
-        thing_to_update.city = request.form['city']
-        thing_to_update.country = request.form['country']
-        thing_to_update.updated_at = datetime.now()
-        try:
-            db.session.commit()
-            flash("Userprofile updated!")
-            return render_template('updateuser.html', form=form, thing_to_update=thing_to_update, id=id)
-        except:
-            flash("Something went wrong :(")
+    user_id = current_user.id
+    if user_id == thing_to_update.id or user_id==19: #adminkammy:
+        if request.method =='POST':
+            thing_to_update.username=request.form['username']
+            thing_to_update.email=request.form['email']
+            thing_to_update.password_hash = generate_password_hash(form.password_hash.data, "sha256")
+            thing_to_update.address = request.form['address']
+            thing_to_update.postal_code = request.form['postal_code']
+            thing_to_update.city = request.form['city']
+            thing_to_update.country = request.form['country']
+            thing_to_update.updated_at = datetime.now()
+            try:
+                db.session.commit()
+                flash("Userprofile updated!")
+                return render_template('updateuser.html', form=form, thing_to_update=thing_to_update, id=id)
+            except:
+                flash("Something went wrong :(")
+                return render_template('updateuser.html', form=form, thing_to_update=thing_to_update, id=id)
+        else:
             return render_template('updateuser.html', form=form, thing_to_update=thing_to_update, id=id)
     else:
-        return render_template('updateuser.html', form=form, thing_to_update=thing_to_update, id=id)
+        flash("You can only edit your own user!")
+        products = db.session.query(Products).all()
+        return render_template('index.html', products=products)
 
-@app.route('/deleteuser/<int:id>')
+@app.route('/deleteuser/<int:id>') #, methods=['DELETE']
 @login_required
 def delete_user(id):
     user_to_delete = Users.query.get_or_404(id)
-
+    user_id = current_user.id
     form = UserForm()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User deleted!")
-        users = db.session.query(Users).all()
-        return render_template('createuser.html', form=form, users=users)
-    except:
-        flash('Something went wrong!')
-        return render_template('createuser.html', form=form, users=users)
+    if user_id == user_to_delete.id or user_id==19: #adminkammy:
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User deleted!")
+            users = db.session.query(Users).all()
+            return render_template('createuser.html', form=form, users=users)
+        except:
+            flash('Something went wrong!')
+            return render_template('createuser.html', form=form, users=users)
+    else:
+        flash('You can only delete your own user!')
+        products = db.session.query(Products).all()
+        return render_template('index.html', products=products)
+
 
 @app.route("/addproduct", methods=['GET','POST'])
 @login_required
@@ -180,39 +196,50 @@ def product_form():
 def update_product(id):
     form = ProductForm()
     thing_to_update= Products.query.get_or_404(id)
+    user_id = current_user.id
+    if user_id == thing_to_update.user.id or user_id==19: #adminkammy
+        if form.validate_on_submit():
+            thing_to_update.name=form.name.data
+            thing_to_update.price=form.price.data
+            thing_to_update.image_url=form.image_url.data
+            thing_to_update.description=form.description.data
+            thing_to_update.updated_at = datetime.now()
 
-    if form.validate_on_submit():
-        thing_to_update.name=form.name.data
-        thing_to_update.price=form.price.data
-        thing_to_update.image_url=form.image_url.data
-        thing_to_update.description=form.description.data
-        thing_to_update.updated_at = datetime.now()
+            #update database
 
-        #update database
-
-        db.session.add(thing_to_update)
-        db.session.commit()
-        flash("Product updated!")
-        return redirect(url_for('show_detail', target_id=thing_to_update.id))
-    form.name.data = thing_to_update.name
-    form.price.data= thing_to_update.price
-    form.image_url.data=thing_to_update.image_url
-    form.description.data=thing_to_update.description
-    return render_template('updateproduct.html', form=form, thing_to_update=thing_to_update)
+            db.session.add(thing_to_update)
+            db.session.commit()
+            flash("Product updated!")
+            return redirect(url_for('show_detail', target_id=thing_to_update.id))
+        form.name.data = thing_to_update.name
+        form.price.data= thing_to_update.price
+        form.image_url.data=thing_to_update.image_url
+        form.description.data=thing_to_update.description
+        return render_template('updateproduct.html', form=form, thing_to_update=thing_to_update)
+    else:
+        flash("Not your product, you can't edit this one!")
+        products = db.session.query(Products).all()
+        return render_template('index.html', products=products)
 
 @app.route('/deleteproduct/<int:id>')
 @login_required
 def delete_product(id):
     product_to_delete = Products.query.get_or_404(id)
+    user_id = current_user.id
+    if user_id == product_to_delete.user.id or user_id==19: #adminkammy:
+        try:
 
-    try:
-        db.session.delete(product_to_delete)
-        db.session.commit()
-        flash("Product deleted!")
+            db.session.delete(product_to_delete)
+            db.session.commit()
+            flash("Product deleted!")
+            products = db.session.query(Products).all()
+            return render_template('index.html', products=products)
+        except:
+            flash('Something went wrong!')
+            return render_template('index.html', products=products)
+    else:
+        flash("Not your product! You can't delete this one!")
         products = db.session.query(Products).all()
-        return render_template('index.html', products=products)
-    except:
-        flash('Something went wrong!')
         return render_template('index.html', products=products)
 
 @app.route("/")
@@ -225,15 +252,23 @@ def index():
 def how_to_solve():
     return render_template('howtosolve.html')
 
+
+#passing stuff to layout html page
+@app.context_processor
+def passthrough():
+    form= SearchForm()
+    return dict(form=form)
+
 @app.route('/search', methods=['POST'])
 def search():
+    form= SearchForm()
+    if form.validate_on_submit():
 
-    if request.method=="POST":
-        searched=request.form['searched']
+        product_searched = form.searched.data
+        products= Products.query.filter(Products.name.ilike('%' + product_searched + '%')).order_by(Products.updated_at).all()
 
-
-        products = db.session.query(Products).filter(Products.name.contains(searched)).all()
-    return render_template('index.html',products=products)
+        return render_template('index.html',form=form, products=products)
+    return render_template('index.html', form=form, products=Products.query.order_by(Products.updated_at))
 
 
 @app.route("/productpage/<string:target_id>")
